+++ app/src/main/java/org/breezyweather/common/basic/wrappers/AirQualityWrapper.kt
@@ -20,7 +20,7 @@ import org.breezyweather.common.basic.models.weather.AirQuality
 import java.util.Date
 
 data class AirQualityWrapper(
-    val current: AirQuality? = null, // TODO: Not supported yet
+    val current: AirQuality? = null,
     val dailyForecast: Map<Date, AirQuality>? = null,
     val hourlyForecast: Map<Date, AirQuality>? = null
 )